- Change ASFSearchResults to subclass of list - ASFProduct no longer subclasses dict
The above two changes allow for a flatter and simpler data structure. results[0] is much easier than results['features'][0]. Additionally, for products, there are now two main properties: .properties, and .geometry. These have all the expected content of the matching fields in a geojson object.

Added str dunders for both ASFSearchResults and ASFProduct. They produce indented, sorted geojson. Along with those, there is now a .geojson() method on both classes that will spit out a dict that matches standard geojson structure (or a geojson feature in the case of ASFProduct)